bugfix: changed fields "sum" in BalanceChange model and fields "active_balance" and "safe_balance" in user model from float to integer; fixed edit_user function in users.views + migrations

+++ app/models.py
@@ -14,7 +14,7 @@ class Category(models.Model):
 
 
 class BalanceChange(models.Model):
-    sum = models.FloatField(verbose_name="Сумма")
+    sum = models.IntegerField(verbose_name="Сумма")
 
     class Type(models.TextChoices):
         breakfast = ('I', 'Income')
